Extract webApp to separate module
