+++ shared/build.gradle.kts
@@ -3,6 +3,7 @@
 import com.mikepenz.aboutlibraries.plugin.DuplicateMode
 import com.mikepenz.aboutlibraries.plugin.DuplicateRule
 import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
+import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
 import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 
 plugins {
@@ -29,26 +30,9 @@ kotlin {
 
     jvm()
 
-    @OptIn(org.jetbrains.kotlin.gradle.ExperimentalWasmDsl::class)
-    wasmJs {
-        binaries.executable()
-        browser {
-            val projectDir = project.projectDir.path
-            val rootDir = project.rootDir.path
-            commonWebpackConfig {
-                outputFileName = "kotlinconf-app.js"
-            }
-        }
-    }
-
-    js {
-        binaries.executable()
-        browser {
-            commonWebpackConfig {
-                outputFileName = "kotlinconf-app.js"
-            }
-        }
-    }
+    @OptIn(ExperimentalWasmDsl::class)
+    wasmJs { browser() }
+    js { browser() }
 
     listOf(
         iosX64(),
