+++ webApp/build.gradle.kts
@@ -1,0 +1,34 @@
+import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
+
+plugins {
+    alias(libs.plugins.kotlinMultiplatform)
+    alias(libs.plugins.composeMultiplatform)
+    alias(libs.plugins.composeCompiler)
+}
+
+kotlin {
+    @OptIn(ExperimentalWasmDsl::class)
+    wasmJs {
+        binaries.executable()
+        browser {
+            commonWebpackConfig {
+                outputFileName = "kotlinconf-app.js"
+            }
+        }
+    }
+
+    js {
+        binaries.executable()
+        browser {
+            commonWebpackConfig {
+                outputFileName = "kotlinconf-app.js"
+            }
+        }
+    }
+
+    sourceSets {
+        commonMain.dependencies {
+            implementation(projects.shared)
+        }
+    }
+}
