+++ webApp/src/webMain/resources/images/KotlinConf2025-icon-360.png
