+++ webApp/src/webMain/resources/images/KotlinConf2025-icon-512.png
