+++ webApp/src/webMain/resources/manifest.json
