+++ webApp/src/webMain/resources/service-worker.js
